캐싱 관련 함수들 utils 폴더에 위치
